Python/zoho_get_contacts/requirements.txt
# Function dependencies, for example:
# package>=version
Flask==2.0.2
requests==2.27.1
google-cloud-bigquery==2.34.4
pandas-gbq==0.16.0